+++ backend/src/test/java/reviewme/review/service/ReviewDetailLookupServiceTest.java
@@ -18,7 +18,7 @@ import reviewme.review.domain.Review;
 import reviewme.review.domain.TextAnswer;
 import reviewme.review.domain.exception.InvalidReviewAccessByReviewGroupException;
 import reviewme.review.domain.exception.ReviewGroupNotFoundByGroupAccessCodeException;
-import reviewme.review.repository.QuestionRepository;
+import reviewme.question.repository.QuestionRepository;
 import reviewme.review.repository.ReviewRepository2;
 import reviewme.review.service.dto.response.detail.TemplateAnswerResponse;
 import reviewme.reviewgroup.domain.ReviewGroup;
