+++ backend/src/test/java/reviewme/review/service/ReviewServiceTest.java
@@ -19,7 +19,7 @@ import reviewme.review.domain.exception.ReviewGroupNotFoundByGroupAccessCodeExce
 import reviewme.review.dto.response.ReceivedReviewCategoryResponse;
 import reviewme.review.dto.response.ReceivedReviewsResponse;
 import reviewme.review.repository.CheckboxAnswerRepository;
-import reviewme.review.repository.QuestionRepository;
+import reviewme.question.repository.QuestionRepository;
 import reviewme.review.repository.Review2Repository;
 import reviewme.reviewgroup.domain.ReviewGroup;
 import reviewme.reviewgroup.repository.ReviewGroupRepository;
